// src/types/store.ts
export interface IStoreData {
  active?: boolean;
  address?: string;
  createDate?: string;
  district?: string;
  email?: string;
  fullName?: string;
  id?: number;
  password?: string;
  phone?: string;
  postcode?: string;
  province?: string;
  storeImage?: string;
  storeName?: string;
  storeType?: string | null;
  subdistrict?: string;
  updateDate?: string;
  userId?: string;
  userLevel?: number;
  userType?: string;
  username?: string;
}